Model for generic oled's
To be used in 3d printed cases as a cutaway hole for generic .96" oled screens (with rounded screw holes).

This model includes a cutaway (the top section) for a piece of .08" plexi to be pressed into place. However, it can be aligned with the surface of the case to whatever depth is preferred.